Fix and refactor `deprecated` decorator to `deprecate_func` (#6594)
* Replace skimage_deprecation with FutureWarning

This was only used in the decorator `deprecated` while everywhere else
FutureWarning is used. So let's be consistent about this.

Also don't modify the warning filter when calling the wrapped function.
That this happens is very implicit and hidden from the user. I'd argue
that it's the users responsibility to configure his warning filter while
it's our responsibility to chose good and expected defaults.

* Make _docstring_add_deprecated private

That function is currently not used anywhere else and very specific to
the deprecate_kwarg decorator. Should probably be part of that class.

* Update decorator for deprecated functions

Previously, the warning did not point to where the deprecated function
was called(user code) but at the deprecated function's definition
(library code)! Add a test for this.

I also propose an updated signature of the decorator, that makes sure
that the deprecated version is specified (forgotten in previous uses
of this decorator). I replaced `alt_func` with a more general `hint`
parameter as there is not always a clear alternative function. In that
case it's possible to provide a more flexible guiding message to the
user.

Also do away with the behavior parameter. Raising an exception defeats
the purpose of a deprecation decorator as it immediately breaks user
code. It impossible to use the deprecated function, in which case it
should just be removed.

* Update current uses of deprecated_func

The deprecation of `compute_hessian_eigenvalues` was introduced after
all 0.19.x releases (see git tag --contains d2689f5)! So the target
should be 0.21.

* Adapt deprecation test of grey* funcs

to updated deprecate_func decorator.

* Apply suggestions from code review

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>
Co-authored-by: Gregory Lee <grlee77@gmail.com>

* Add docstring example to deprecate_func

* Update old reference to new deprecate_func

* Apply suggestions from code review

Co-authored-by: Jarrod Millman <jarrod.millman@gmail.com>

---------

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>
Co-authored-by: Gregory Lee <grlee77@gmail.com>
Co-authored-by: Jarrod Millman <jarrod.millman@gmail.com>