Create data_dir only on actual download
Even though skimage.data uses lazy loading, its submodule `_fetcher.py`
is executed when `skimage` is imported, because its attribute `data_dir`
is imported in multiple places [1]. Previously, this lead to
`_init_pooch()` being always executed, which in turn tried to create the
data directory preemptively. This lead to problems when `data_dir`
wasn't writeable, e.g. when scikit-image is used in read-only
containers.

This refactoring alleviates this by postponing the directory creation
until it is actually needed and data is being downloaded. Calling
`download_all()` also ensures that legacy files are copied to
`data_dir`; this use case was requested in [2] and should be preserved
this way.

With the previous and current state, the behavior is somewhat difficult
to test with regard to self-contained tests and multi-processing /
multi-threading. `_fetchers.py` should probably be refactored into a
class whose API is less intertwined with global state; that might be
easier to test.

[1] https://github.com/scikit-image/scikit-image/blob/51225d16ddebffacd0ccd9a54d06a673e3caff98/skimage/__init__.py#L141
[2] https://github.com/scikit-image/scikit-image/pull/3945#issuecomment-498141893