Remove copy_legacy_to_cache flag in _fetch
Instead, it is now the task of `download_all(directory=...)`` to place a
copy of every data_file in `directory` or - if not given - the default
cache directory.

This also addresses another previously undiscovered bug. Running

    import skimage as ski
    ski.data.download_all()
    ski.data.download_all(directory="example_dir")

would not create anything in "example_dir" because `_fetch` would always
return the cached entry before ever invoking pooches cache mechanism to
place it in "example_dir".