Use legacy datasets without creating a `data_dir` (#6886)
* Replace os.path with pathlib

* Create data_dir only on actual download

Even though skimage.data uses lazy loading, its submodule `_fetcher.py`
is executed when `skimage` is imported, because its attribute `data_dir`
is imported in multiple places [1]. Previously, this lead to
`_init_pooch()` being always executed, which in turn tried to create the
data directory preemptively. This lead to problems when `data_dir`
wasn't writeable, e.g. when scikit-image is used in read-only
containers.

This refactoring alleviates this by postponing the directory creation
until it is actually needed and data is being downloaded. Calling
`download_all()` also ensures that legacy files are copied to
`data_dir`; this use case was requested in [2] and should be preserved
this way.

With the previous and current state, the behavior is somewhat difficult
to test with regard to self-contained tests and multi-processing /
multi-threading. `_fetchers.py` should probably be refactored into a
class whose API is less intertwined with global state; that might be
easier to test.

[1] https://github.com/scikit-image/scikit-image/blob/51225d16ddebffacd0ccd9a54d06a673e3caff98/skimage/__init__.py#L141
[2] https://github.com/scikit-image/scikit-image/pull/3945#issuecomment-498141893

* Fix typo in docstring

* Fix test errors due to pathlib refactoring

Also, examples in `ImageCollection`'s docstring [1] indicate that this
is part of our API, so revert `data_dir` back to being a string.

* Fix return types

* Revert refactoring to pathlib

Long-term, I would think of an update from `os.path` to `pathlib` as
removing technical debt but it was a bad call in the context of this
fix. Especially, because it had unintended side-effects to our API.

* Ensure cache subdir exists

* Debug failure on macos-cp3.11

* Debug: Use absolute paths

* Debug: ignore legacy path

multipage_rgb.tif is not in our distribution archives.

* Debug: remove codecov dependency

Super strange, but suddenly it seems that codecov has disappeared from
PyPI [1]...

[1] https://pypi.org/project/codecov/

* Debug: use --showlocal for pytest

* Debug: test hashes and pure imread

* Remove missed codecov in pyproject.toml

* Debug check sorting

* Always try cache first when fetching datasets

* Remove debug test

* Remove copy_legacy_to_cache flag in _fetch

Instead, it is now the task of `download_all(directory=...)`` to place a
copy of every data_file in `directory` or - if not given - the default
cache directory.

This also addresses another previously undiscovered bug. Running

    import skimage as ski
    ski.data.download_all()
    ski.data.download_all(directory="example_dir")

would not create anything in "example_dir" because `_fetch` would always
return the cached entry before ever invoking pooches cache mechanism to
place it in "example_dir".

* Use proper cache_dir in _fetch without pooch too

* Expand user in download_all

Previously, running

    import skimage as ski
    ski.data.download_all("~/skimage-data")

would place files at two locations: files in the distribution are placed
in "[working_dir]/~/skimage-data" while files downloaded with pooch were
placed in /home/[user]/skimage-data. I think this was because our old
os.path machinery doesn't resolve ~ while pooch uses pathlib which does.

To address this we make download_all explicitly expand the user if
directory is given.

---------

Co-authored-by: Stefan van der Walt <stefanv@berkeley.edu>