Use not-palleted version of eagle.png
Updates the registry entry for eagle.png to the updated version from
[1]. The original file contained a RGB palette whose contents were
identical. The image is therefore effectively grayscale. Previous to
imageio v2.28, their pillow plugin had custom logic to check for this
case in which case it would remove the duplicate channels. However, this
behavior was removed in v2.28, which means that reading the image will
return an array with those RGB channels as the last dimension which
breaks scikit-image's existing API and tests. [2]

[1] https://gitlab.com/scikit-image/data/-/merge_requests/22
[2] https://github.com/scikit-image/scikit-image/pull/6764#issuecomment-1502737516