Rework docstring of ImageCollection
Removes references to imageio.get_loader whose use is not recommended
this way. The example demonstrates a pattern where it is easy to
leave the file in an unclosed state. Considering the uncertain state
of skimage.io.imread and imageio being an external library, I chose
not to educate users here on its usage. Users might be better of looking
at imageio's documentation or dedicated places in our docs.

Furthermore, the distinction of the notes section from the the example
section wasn't really clear. This moves the code snippets into the
example section.