Skip unstable test in `ransac`'s docstring
I'm seeing this test fail sometimes in our CI. It's creating noise that confuses contributors and is annoying to us. I'd prefer to pin the test's instead. But until there is a decision on pinning our randomness we should skip this line. Most other output in this docstring seems to be intentionally untested as well.