apply spacing rescaling when computing centroid_weighted (#6900)
* apply spacing rescaling when computing centroid_weighted

the slice.start coordinates are still in pixel and not yet rescaled with
the given spacing and should be rescaled before adding them to the
coordinate of the local centroid_weighted

* add more test coverage to test_centroid_weighted when using spacing

the default SAMPLE has a bounding box as big as the original image, so
the slice.start is in (0, 0). When using spacing the slice.start needs
to be rescaled as well, but having it in the origin hides potential
errors in doing so. An image with the prop bounding box away from the
origin can be used instead