Expose `GeometricTransform.residuals` in HTML doc (#6968)
`GeometricTransform` as a class itself is not part of the official API
and documentation. So let's make that apparent by prefixing with "_" and
making it a proper abstract base class so that it can't be instantiated
and subclasses raise an error if they don't implement an abstract
method.

Using the `:inherited-members:` directive makes sure that the method
`_GeometricTransform.residuals` and its docstring show up in our HTML
doc and makes the interface of each class clearer without having to jump
around.