Fix sparse index type casting
In SciPy 0.12, we are no longer guaranteed that indices will be 32-bit
in most situation. In fact, that was never a valid assumption, but one
that often worked out in practice.

The approach taken here is not ideal: all indices are cast to
64-bit, because we don't *know* whether 32 or 64-bit indices are being
used.

The issue was exposed by https://github.com/scipy/scipy/pull/18509,
which no longer down-casts indices, once they've been upcast.