Rework docstring of ImageCollection
Removes the usage of __file__ from the example (will not work when
copying the example to console). Furthermore, the distinction of the
notes section from the the example section wasn't really clear.
This merges the code snippets into the example section and hopefully
simplifies the docstring.

Also the load_func example no longer relies on no_time_for_that_tiny.gif
which is a good thing in my book. It's not really a part of our public
API and has made problems in the past.