Refactor generate_release_notes.py
This refactor is supposed to make the script re-usable and easier to
maintain in anticipation of moving it to the umbrella of the
scientific-python organisation.

Also changes / improvements:

Access to and parsing of a local git repository is no longer necessary.
Instead, the github.Repository.compare() is used to retrieve all commits
of interest.

Pull requests are no longer retrieved from commits by parsing the commit
message with a regex. Instead, github.Commit.get_pulls() is used to
retrieve them.

RestructuredText and Markdown are now both supported.

In case the notes are formatted in RST, single backticks "`" are
transformed accordingly to "``" (using naive replacement).