Add co-authors list
For commit authors we get a login handle (nice unique key) and name from
GitHub's API. For co-authors, I couldn't find a way to get them via the
API.

So I fell back to parsing the commit message which only yields a name
and email pair. Co-authors are considered unique by their parsed name,
and a random URL to a co-authored commit is included to make them more
discoverable.