Add bin splitting to rank.percentile_mean
The algorithm works by summing bins of a histogram of the surrounding
neighborhood together while skipping bins outside the defined
percentile. However, the previous implementation would only deal with
complete bins leading to unexpected behavior when according to the
given percentile, a part of a bin should factor into the mean. The
new implementation fixes this problem, while increasing performance such
that it is on par with rank.mean for most tested scenarios.