Update pytest config in `pyproject.toml` with repo-review recommendations (#7063)
Added required "ini_options" to pytest config in pyproject.toml. The field is required according to [1]. From local testing it seemed like these options weren't picked up by pytest.

The benchmark suite is currently broken. The "testpaths" option is to be uncommented once the suite is fixed (in another PR).

[1] https://docs.pytest.org/en/7.1.x/reference/customize.html#pyproject-toml
[2] https://learn.scientific-python.org/development/guides/repo-review/

---------

Co-authored-by: Jarrod Millman <jarrod.millman@gmail.com>