Make PyWavelets an optional dependency
denoise_wavelet actually works without PyWavelets for the 1D case if
`convert2ycbr` is True and `estimate_sigma` is False.

Use xfail instead of pytest.importorskip or pytest.mark.skip, so that
the above behavior can actually be tested and we notice if the failure
is due something different (not an ImportError).