Deprecate return_error only in 0.23
Until 0.22 we told users to pass return_error='always' to this function
to silence the warning. Starting with 0.22 this will immediately return
an error without any explanation.

I kind of feel like removing the parameter completely requires another
deprecation? It get that this is totally annoying but that's how I would
interpret our deprecation guide lines and the user experience.