Remove unused CIBW_BUILD env variables
When using reusable workflows, in this case `wheels_recipe.yml`,
environment variables defined in the calling workflow aren't passed [1].
Currently these are already defined with the same values in
`wheels_recipe.yml` so remove them. If a custom value is needed in the
future, these should be added as inputs.

[1] https://docs.github.com/en/actions/using-workflows/reusing-workflows#limitations