Test building nightly wheels with nightly NumPy 2.0 (#7251)
* Test building nightly wheels with nightly NumPy 2.0

Building with NumPy 2.0 should also be compatible with NumPy 1.x at
runtime. However, the other way around doesn't work.

Added a trigger on push to `maintenance/**` prefixed branches (like the
`wheels_test_and_release.yml` job). This should build the nightly wheels
of scikit-image but not upload them to Anaconda.

When using reusable workflows, in this case `wheels_recipe.yml`,
environment variables defined in the calling workflow aren't passed [1].
Currently these are already defined with the same values in
`wheels_recipe.yml` so remove them. If a custom value is needed in the
future, these should be added as inputs.

[1] https://docs.github.com/en/actions/using-workflows/reusing-workflows#limitations