Handle floating point warning for empty images in phase_cross_correlation
Starting with pytest 8.0.0rc1, the following warning triggered the test
in question to fail:

RuntimeWarning: invalid value encountered in scalar divide

For earlier pytest versions, this warning was silently consumed by
pytest.warns (bad!). Now, it's actually showing.

Replace this warning with a more expressive one and test for it as well.