Use pytest-doctestplus
This has the advantage of being able to hide custom test setup from
the rendered HTML docstrings and many more perks we might use in the
future.

doctestplus has its own mechanism of skipping doctests if dependencies
are not required. However, it seems that using `__doctest_skip__` will
not show that tests were skipped, so instead I'm using an approach
based on `.. testsetup` and `pytest.importorskip`.