Deprecate and vendor lookfor
Unfortunately we relied on NumPy's lookfor which is promptly removed in
NumPy 2.0. Technically, if we don't want to make a breaking release we
need to deprecate the function in a deprecation cycle for which we need
to vendor NumPy's lookfor.

Though, skimage.lookfor is an interactive function, so we might get away
with not vendoring NumPy's lookfor, and just making it return a
deprecation warning...