Make it possible to run tests without numpydoc
Skip the three tests requiring numpydoc when it is not installed.
For `skimage/_shared/tests/test_utils.py`, use the existing
`have_numpydoc` constant.
For `skimage/measure/tests/test_regionprops.py`, use the standard
`pytest.importorskip()` function.

This makes it possible to successfully run the test suite when
`numpydoc` is not installed.  This would be convenient for Gentoo
since numpydoc is facing neverending test regressions (in its own test
suite).