Remove obsolete np.set_printoptions
This option should have no bearing on the precision with which results
are compared by `assert_array_almost_equal`. The parameter decimal=10
should be used for that purpose.

This option also had unwanted side effects, because it changed the
display global option globaly leading to problems in later doctest
comparisons.

Note that changing the comparisons in this test to use
`assert_array_almost_equal(..., decimal=10)` fails the tests.