Revert to Python float that is out of range
In `test_overrange_tolerance_float`:
The previous test actually checked what happens if a Python float that
is out of range for the image type is used as the tolerance. Not casting
with `max_value.item()` would result in `np.float32(inf)` which is a
different test.

In `flood`:
Furthermore, we need to turn `min_value` and `max_value` into Python
scalars with `.item()` as well. Otherwise `max` and `min` seem to
trigger an overflow error during the comparison.