Test that skeletonize doesn't modify original input
Due to skeletonize wrapping two algorithms handling and copying the
input is all over the place. In general, I'm trying to reduce the amount
of unnecessary copies while being tolerant with user input (dtype &
memory layout) and keeping the diff reasonable.

So a test seems like the best way to ensure that it stays that way.