Deprecate `skimage.morphology.skeletonize_3d` in favor of just `skeletonize` (#7094)
* Include skeletonize_2d in our public API and doc

* Deprecate skeletonize_3d in favor of skeletonize

instead of making skeletonize_2d public.

* Remove scaling in skeletonize_3d

Previously, skeletonize_3d returned a ubyte array whose values were
either 0 (background) or maxvalue (skeleton). maxvalue was determined
by scaling with img_as_ubyte and then taking the max. I don't really see
a value in this, the skeleton still has the same value everywhere, e.g.
labels with different values are not treated as separate objects /
skeletons. It just lead to unexpected behavior in [1].

[1] https://github.com/scikit-image/scikit-image/issues/7090

* Test that skeletonize doesn't modify original input

Due to skeletonize wrapping two algorithms handling and copying the
input is all over the place. In general, I'm trying to reduce the amount
of unnecessary copies while being tolerant with user input (dtype &
memory layout) and keeping the diff reasonable.

So a test seems like the best way to ensure that it stays that way.

* Remove unnecessary deprecation of skeletonize_2d

That function wasn't public before so we don't need to do a deprecation
cycle for it.

* Use consistend parameter description

across all skeletonize* functions.

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>

---------

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>