Address new copy semantics & broadcasting in `np.solve` in NumPy 2 (#7341)
* Use NP_COPY_IF_NEEDED for compatibility with new copy semantics

With NumPy 2.0.0, `copy=False` now raises a ValueError if the copy
cannot be made. The previous behavior to only copy if needed is provided
with `copy=None`. During the transition period, use this symbol instead.
Remove once NumPy 2.0.0 is the minimal required version [1, 2]

This approach was inspired by astropy [3].

[1] https://numpy.org/devdocs/release/2.0.0-notes.html#new-copy-keyword-meaning-for-array-and-asarray-constructors
[2] https://www.github.com/numpy/numpy/pull/25168
[3] https://github.com/astropy/astropy/blob/2390bf8f8f2a1c5b6beffbd87db9af5aa6538df1/astropy/utils/compat/numpycompat.py#L29

* Address new broadcasting rules in NumPy 2's `solve`

"The broadcasting rules for `np.solve(a, b)`` were ambiguous when b had 1
fewer dimensions than `a`. This has been resolved in a backward-
incompatible way and is now compliant with the Array API. The old
behaviour can be reconstructed by using
`np.solve(a, b[..., None])[..., 0]`."

https://numpy.org/devdocs/release/2.0.0-notes.html#removed-ambiguity-when-broadcasting-in-np-solve
https://www.github.com/numpy/numpy/pull/25914