Deprecate output and positional args in `filters.gaussian` (#7225)
* Deprecate output and positional args in filters.gaussian

Addresses an item in our skimage2 transition list [1]. We can't simply
use deprecate_kwarg or remove_arg, as `output` can also be passed as a
positional argument. Instead, deprecate passing more than 1 positional
argument and at the same time, refactor `output` to `out`.

I'm not really sure whether remove_arg is a user friendly way to
deprecate a positional parameter. Applying the function prevents
positional arguments after the deprecated one as well, even though this
is not deprecated. The warning will users also not tell them about it.
I find it clearer to just deprecate these positional arguments in favor
of keyword-only ones. I also think it's the better practice.

[1] https://github.com/scikit-image/scikit-image/wiki/API-changes-for-skimage2

* Use deprecate_parameter

* Tweak docstring and error message

according to code review.

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>

* Set deprecation stop date according to guidelines

According to [1], deprecations should be inplace for two releases. If
merged this will be releases 0.23 and 0.24. So 0.25 will finalize the
deprecation.

[1] https://scikit-image.org/docs/dev/development/contribute.html#deprecation-cycle

* Apply suggestions from Marianne

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>

* Tweak error message for out

as suggested by Marianne. I felt like "got" conveyed the same meaning
as "provided" but is a bit less advanced vocabulary.

* Use assert_stacklevel

which was recently merged.

---------

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>