Simplify instructions about changelist in PR template (#7401)
I noticed that several contributors struggled with interpreting these
instructions properly. So rather than making them attempt something they
are not familiar with, let's make it clear that we don't expect them to
fill this out.

Also remove the newlines. It seems that GitHub now (?) preserves
newlines even within paragraphs. Can't find any news on this but I'm
sure that this wasn't the previous behavior...

Co-authored-by: Jarrod Millman <jarrod.millman@gmail.com>