Add function to remove near objects in nd-image (#4165)
* Move shared nd-utility functions into _util module

* Add remove_close_objects function

This function iterates over all objects (connected pixels that are True)
inside an image and removes neighboring objects until all remaining ones
are at least a minimal euclidean distance from each other.

* Test basic functionality of remove_close_objects

* Add manual bounds checking

* Improve tests and handle edge cases better

* Support boolean views on numeric dtypes

* Simplify Cythonization in morphology.setup.py

_extrema_cy.pyx was cythonized twice for now reason.

* Sort objects by their size if no priority is given

* Add support for int and float type images

* Add test for object with unequal values

* Simplify/tweak helper functions in morphology._util

Simplify construction of indices, use consistent name for structuring
element (selem), some reformatting and improve documentation.

* Only evaluate object surface in remove_close_objects

This trick significantly improves the performance by reducing the number
of points inside the KDTree whose neighborhood needs to be evaluated
for objects that are to close. The greater the checked objects size to
surface ratio, the greater the improvement.

* Use intp as label dtype

numpy.bincount seems to cast its input to dtype('intp') according to
the rule safe. Therefore giving a dtype such as uint32 or int64 that
can hold a larger number fails on 32-bit platforms.

* Add basic benchmarks for remove_close_objects

* Correct faulty results in docstring example

With recent changes the evalutation order of objects changed.

* Use "mergesort" instead of "stable" sort

The latter was added in NumPy 1.15 which is not yet a minimal
requirement for scikit-image.

* Use objects' size as priority in benchmark

* Document behavior for NaNs and provide test

* Support raveled offsets for dimensions smaller selem

The old implementation would raise a ValueError if a dimension length
given in `image_shape` was smaller than in `selem`. This precluded
operations on images e.g. with the shape (2, 100).

* Pass-through p-norm to KDTree

This allows the user to control how the distance between objects/samples
is calculated (e.g. Euclidean, Manhatten, Chebyshev).

* Address unintended reorering by np.unique

* Remove parameter not supported by Scipy 0.19

return_sorted was added in SciPy 1.2 and is not yet supported by
scikit-images minimal required version.

* Reorder parameters in remove_close_objects

* Rename to remove_near_objects

"close" can have multiple meanings especially in the context of
morphological operations (closing). So the synonym "near" should have
less potential for confusion.

* Use in_place instead of inplace

* Add remove_near_objects to API and release notes

* Don't class-bundle remove_near_objects's tests

* Assert Exception for mismatching priority shape

* Fix whitespace

* Add gallery example for removing objects

This example demonstrates the usage of remove_small_objects and
remove_near_objects.

* Fix doctest in remove_near_objects

* Improve misleading doc wording

* Use shared fused dtype np_real_numeric

Already declared, so no need to duplicate it here.

* Use out argument instead of in_place

and rename selem to footprint for a more consistent API.

ravel is replaced with reshape(-1) as recommended by NumPy's doc.

* Group test's for remove_near_objects in class

It's easier to seperate the tests for this functions from the other
ones this way.

* Simplify input sanitation

_offsets_to_raveled_neighbors is now a just a wrapper for
_raveled_offsets_and_distances, so replace it.

Special handling of an empty `image` is no longer necessary as both
_raveled_offsets_and_distances and cKDTree support arrays with size 0.

* Handle memory layout more expicitely

`out` must be C-contiguous which allows for efficient indexing in the
underlying Cython code. This might raise an error if out is supplied by
the user so document it with a test.

It seems that special logic for bool-type arrays is no longer necessary
as the appropriate PR [1] was merged.

[1] https://github.com/cython/cython/pull/2676

* Remove trailing whitespace

to make CI linter happy.

* Add _near_objects_cy to new meson system

and fix import path of _shared.fused_numerics. Of course wouldn't
compile otherwise.

* Address NumPy deprecation warning

The warning was found by the CI (linux-cp3.9-pre) for NumPy v1.24.1:

DeprecationWarning: NumPy will stop allowing conversion of out-of-bound
Python integers to integer arrays.  The conversion of -1 to uint32 will
fail in the future.

The objects don't have to be -1, just any number != 0.

* Add remove_near_objects to morpholoy.__all__

Maybe missed during merging?

* Use labels as input and map priority to index

This overhauls the API of remove_near_objects. Instead of labeling the
input itself, the function already expects a labeled image. Sorting an
index into this label_image by label ID first, and by priority second,
makes it easy to find other indices belonging to the same object because
they are contiguous in the resulting index. This makes it possible to
entirely omit the parameters connectivity and footprint (for now).

`priority` now is expected to assign a priority to each object and no
longer gives a priority for each pixel.

Test and benchmarks were updated accordingly.

* Ignore labels which aren't surface

* Add two new test cases one of which currently fails

Also sadly the optimization isn't working as expected. See
test_objects_with_inside.

* Allow F-contiguous arrays

requires an additional copy but still, good to support.

* Deal with constant and negative input

* Remove obsolete _max_priority

* Release GIL in _remove_object

* Update docstrings

* Format with black

* Don't use boolean input in gallery example

* Use C++ set and release GIL

* Exclude inner samples of objects from KDTree & critical loop

Even with the additional cost of separating bounday and inner samples
of objects, this improves the performance significantly!

* Improve documentation of remove_near_objects

* Color labels in gallery example

* Revert to current version of 0.20 release notes

* Add support for anistropic data in remove_near_objects

and also add a random test that uses CKDTree.sparse_distance_matrix to
manually check distance between remaining objects.

* Rename to min_distance like in peak_local_max

or corner_peaks.

* Test p-norm as well

* Correct spelling

* Use sparser borders if possible

for unity spacing and if p-norm is 1 or 2. Also refactor code structure
a little bit.

* Apply suggestions from code review

Co-authored-by: Stefan van der Walt <sjvdwalt@gmail.com>

* Tweak docs based on Stéfan's suggestions

Co-authored-by: Stefan van der Walt <sjvdwalt@gmail.com>

* Refactor to remove_objects_by_distance

* Handle np.bincount expecting intp on WASM

Tried to handle this in a more general sense. Not sure if this is the
best option though.

* Use is_wasm to handle np.bincount expecting intp

* Apply suggestions from code review

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>

---------

Co-authored-by: Stefan van der Walt <sjvdwalt@gmail.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>