Fix CI tests with minimal dependencies and make dependency resolution more robust (#7462)
* Try installing default deps after build

NumPy 2 is minimal dep for build but not for default. So installing
default before building actually downgrades NumPy for the build step.
This becomes a problem later when testing.

* Remove unused cruft from before_install.sh

* Escape $filename with double quotes

* Include previous requirements in successive pip installs

Installing successively may update previously constrained dependencies,
because pip apparently forgets these constraints. This way the
pinned minimal dependencies should be kept or at least we should get
a resolution error with conflicting constraints.

* Fix typo
Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>