Fixup conda development installation instructions
I believe that these two minor fixes should be worth our effort in order
to maintain compatibility with conda.

The first seems to be a quirk of pip and pypi which alias underscores
and dashes. This isn't true in conda(-forge) which requires two separate
packages.

The second one line change simply updates the command to use conda-forge
for all build packages since they are now available there.