Improve robustness of _dynamic_max_trials
For a larger number of minimum samples the denominator is equal to one
due to rounding errors which gives a logarithm of zero. Since the
logarithm of the nominator is negative, we get as a return value -np.inf.

To prevent the rounding errors, limit the nominator to a number slightly
smaller than one.

As a robustness measure, the nominator is also being clipped to make
sure we always get a positive number for the maximum number of trials.