Uncomment currentmodule directive again
This line was left commented as an oversight from debugging
https://www.github.com/scikit-image/scikit-image/issues/7485 and
https://www.github.com/scikit-image/scikit-image/pull/7486.

It didn't seem to cause any problems but let's put it back in to be
safe. I'm a bit fuzzy on this because I can't find a good answer in
Sphinx docs to how those two directives interact and why it seems to
work with and without.