Use pytest config in pyroject.toml in CI
This should be safe, as the pytest documentation [1] states that
"rootdir is NOT used to modify sys.path/PYTHONPATH". So while the
`--config-file` option affects the `rootdir`, it's not adding that
directory to the PYTHONPATH where it could accidentally overwrite the
installed package. This solution is somewhat unsatisfying. Personally,
I think a better approach would be to switch to the src-layout.

[1] https://docs.pytest.org/en/7.1.x/reference/customize.html?highlight=rootdir#initialization-determining-rootdir-and-configfile