Preserve value of deprecated parameters (#7552)
Even if a parameter is deprecated, a value passed to it must be
preserved during the deprecation cycle! Othwerwise, we effectively
remove the deprecated behavior before the cycle is complete.

This is only relevant if there's no replacement for the deprecated
parameter. In that case, the given value preserved and passed to the
replacing parameter, and the deprecated one can be overwritten with
`DEPRECATED`.