Ensure that RegionProperties can be deserialized with pickle (#7569)
When deserializing a RegionProperties object with pickle, `__setstate__`
[1] is accessed before any other attributes like are available. This
leads to an RecursionError in `__getattr__` which doesn't account for
this and tries to call attributes such as `self._intensity_image`. We
can guard against this in multiple ways. I chose a very specific one in
this case which makes sure that the default AttributeError is provoked
(gh-6465).

[1] https://docs.python.org/3/library/pickle.html#object.__setstate__