Build sphinx documentation with parallel jobs (#7579)
* Build sphinx documentation with parallel jobs

Starting with v0.17, sphinx-gallery supports generating galleries in
parallel. For our setup I see a reduction of runtime from ~9 min to
~6 min when run 8 parallel jobs. I think our GitHub runners have 3-4
cores, which should still give a nice improvement.

Locally I see a UserWarning:
A worker stopped while some jobs were given to the executor. This can be
caused by a too short worker timeout or by a memory leak.

Not sure yet what is causing this or if it is something to worry about.

* Use `-j auto` on Azure

* Require joblib indirectly via dependency group of sphinx-gallery and
clarify comment.

---------

Co-authored-by: Jarrod Millman <jarrod.millman@gmail.com>