Deprecate `skimage.io` plugin infrastructure (#7353)
* Deprecate plugin infrastructure in skimage.io

but keep imread, imsave and imread_collection around to work as simple
wrappers for imageio down the line.

* Preserve value of deprecated parameters

Even if a parameter is deprecated, a value passed to it must be
preserved during the deprecation cycle! Othwerwise, we effectively
remove the deprecated behavior before the cycle is complete.

This is only relevant if there's no replacement for the deprecated
parameter. In that case, the given value preserved and passed to the
replacing parameter, and the deprecated one can be overwritten with
`DEPRECATED`.

* Remove deprecated io functions from __all__

* Try if `append=True` works for all tifffile versions

we are looking for something that we can pass to `**plugin_args` to
check whether a deprecation warning is raised.

* Note replacing call_plugin with imageio plugin in TODO

* Reword plugin deprecation message

---------

Co-authored-by: Jarrod Millman <jarrod.millman@gmail.com>