Feature/enhance plot matched features (#7541)
* Fix seed for deterministic colors

* Allow sequence of colors for plot_matched_features

* Enhance unit test for plot_matched_features

* Avoid error message duplication

* Simplify logic

Checking isinstance of Sequence doesn't allow for arrays of correct shape. So
instead, just check for the valid length.

* Don't use zip in iteration

`zip(matches, colors)` doesn't fail if both arguments have a different
length and just drops items silently. Make it fail, in case we ever mess
up something before the iteration.

* Add tests for exceptions

* Move `rng` closer to where it's used

* Skip tests if matplotlib is not installed

Whoops.

* Update skimage/feature/util.py

Improve parameter description of `matches_color`

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>

* Update skimage/feature/util.py

Update error message

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>

* Improve wording of comments

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>

* Add reference to supported color formats

* Check for complete error message

Checking for an error message that `matches_color` needs to be a single color is confusing
because we support also sequences.

* Enhance test cases and remove unused variable

* Pin randomness in `test_plot_matched_features`

Also, initializing the tested value into a named variable before passing
it to `matches_color` should simplify it somewhat and makes the code
easier to debug.

* Note use of itertools.cycle for color map

and make it a bit clearer that `matches_color` needs to correspond in
lenght to `matches` if not a single color.

* Use Random Generator with seed instead of old API

* Reword note on using itertools.cycle

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>

---------

Co-authored-by: Lars Grüter <lagru@mailbox.org>
Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>
Co-authored-by: Jarrod Millman <jarrod.millman@gmail.com>