Build docs in parallel instead of sequentially on Azure
This should speed-up our wait time for the CI to finish completely
significantly, since pytest + doc building on Azure is the longest
job in our current CI matrix for PRs. This was discussed during a
community meeting (2024-10-22).

Also do away with testing on Python 3.10, which is no longer within
scope of SPEC 0.